Generalized the numpy2message stuff
